kod.13.5.task: tips have benn taken, almost all
