+++ kodilla-good-patterns/src/main/java/flightSearcher/Application.java
@@ -6,11 +6,11 @@ public class Application {
     public static void main(String[] args) {
 
         Flight flight = new Flight("Warszawa", "Londyn");
-        FlightProcessor flightProcessor = new FlightProcessor(new CreateFlightList(), new FindFlight(flight));
+        FlightProcessor flightProcessor = new FlightProcessor();
 
-        List<Flight> flightsFrom = flightProcessor.findFlightFrom(flight);
-        List<Flight> flightsTo = flightProcessor.findFlightTo(flight);
-        List<FlightPair> indirectFlights = flightProcessor.findIndirect();
+        List<Flight> flightsFrom = flightProcessor.findFlightFrom(flight.getFrom());
+        List<Flight> flightsTo = flightProcessor.findFlightTo(flight.getTo());
+        List<FlightPair> indirectFlights = flightProcessor.findIndirect(flight);
 
         System.out.println("All flights from: " + flight.getFrom() + " are: " + flightsFrom + ".");
         System.out.println("All flights to " + flight.getTo() + " are: " + flightsTo + ".");
