+++ kodilla-good-patterns/src/main/java/flightSearcher/Flight.java
@@ -1,5 +1,6 @@
 package flightSearcher;
 
+import java.util.List;
 import java.util.Objects;
 
 public class Flight {
